requirements.txt
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
et_xmlfile==2.0.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
numpy==2.0.2
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
xlrd==2.0.1
